rename pg -> bq
